training/requirements.txt
numpy==1.23.5
Pillow==10.0.1
Pillow==10.1.0
streamlit==1.27.2
tensorflow==2.14.0
tensorflow_intel==2.14.0
